+++ packages/swirl-components/src/components.d.ts
@@ -50,6 +50,7 @@ import { SwirlPopoverAnimation } from "./components/swirl-popover/swirl-popover"
 import { ComputePositionReturn, Placement, Strategy } from "@floating-ui/dom";
 import { SwirlProgressIndicatorSize, SwirlProgressIndicatorVariant } from "./components/swirl-progress-indicator/swirl-progress-indicator";
 import { SwirlRadioState, SwirlRadioVariant } from "./components/swirl-radio/swirl-radio";
+import { SwirlBoxPadding as SwirlBoxPadding1 } from "./components/swirl-resource-list/swirl-resource-list";
 import { SwirlResourceListItemLabelWeight } from "./components/swirl-resource-list-item/swirl-resource-list-item";
 import { SwirlSearchVariant } from "./components/swirl-search/swirl-search";
 import { SwirlSeparatorColor, SwirlSeparatorOrientation, SwirlSeparatorSpacing as SwirlSeparatorSpacing1 } from "./components/swirl-separator/swirl-separator";
@@ -63,6 +64,7 @@ import { SwirlSwitchLabelPosition } from "./components/swirl-switch/swirl-switch
 import { SwirlSymbolSize } from "./components/swirl-symbol/swirl-symbol.types";
 import { SwirlTabPadding } from "./components/swirl-tab/swirl-tab";
 import { SwirlTabBarJustify, SwirlTabBarPadding, SwirlTabBarTab, SwirlTabBarVariant } from "./components/swirl-tab-bar/swirl-tab-bar";
+import { SwirlTableDropRowEvent } from "./components/swirl-table/swirl-table";
 import { SwirlTableColumnSort } from "./components/swirl-table-column/swirl-table-column";
 import { SwirlTabBarJustify as SwirlTabBarJustify1, SwirlTabBarPadding as SwirlTabBarPadding1, SwirlTabBarVariant as SwirlTabBarVariant1 } from "./components/swirl-tab-bar/swirl-tab-bar";
 import { SwirlTagIconPosition, SwirlTagIntent, SwirlTagSize, SwirlTagVariant } from "./components/swirl-tag/swirl-tag";
@@ -122,6 +124,7 @@ export { SwirlPopoverAnimation } from "./components/swirl-popover/swirl-popover"
 export { ComputePositionReturn, Placement, Strategy } from "@floating-ui/dom";
 export { SwirlProgressIndicatorSize, SwirlProgressIndicatorVariant } from "./components/swirl-progress-indicator/swirl-progress-indicator";
 export { SwirlRadioState, SwirlRadioVariant } from "./components/swirl-radio/swirl-radio";
+export { SwirlBoxPadding as SwirlBoxPadding1 } from "./components/swirl-resource-list/swirl-resource-list";
 export { SwirlResourceListItemLabelWeight } from "./components/swirl-resource-list-item/swirl-resource-list-item";
 export { SwirlSearchVariant } from "./components/swirl-search/swirl-search";
 export { SwirlSeparatorColor, SwirlSeparatorOrientation, SwirlSeparatorSpacing as SwirlSeparatorSpacing1 } from "./components/swirl-separator/swirl-separator";
@@ -135,6 +138,7 @@ export { SwirlSwitchLabelPosition } from "./components/swirl-switch/swirl-switch
 export { SwirlSymbolSize } from "./components/swirl-symbol/swirl-symbol.types";
 export { SwirlTabPadding } from "./components/swirl-tab/swirl-tab";
 export { SwirlTabBarJustify, SwirlTabBarPadding, SwirlTabBarTab, SwirlTabBarVariant } from "./components/swirl-tab-bar/swirl-tab-bar";
+export { SwirlTableDropRowEvent } from "./components/swirl-table/swirl-table";
 export { SwirlTableColumnSort } from "./components/swirl-table-column/swirl-table-column";
 export { SwirlTabBarJustify as SwirlTabBarJustify1, SwirlTabBarPadding as SwirlTabBarPadding1, SwirlTabBarVariant as SwirlTabBarVariant1 } from "./components/swirl-tab-bar/swirl-tab-bar";
 export { SwirlTagIconPosition, SwirlTagIntent, SwirlTagSize, SwirlTagVariant } from "./components/swirl-tag/swirl-tag";
@@ -1847,6 +1851,11 @@ export namespace Components {
         "assistiveTextItemMoving"?: string;
         "controllingElement"?: HTMLElement;
         "label"?: string;
+        "padding"?: SwirlBoxPadding1;
+        "paddingBlockEnd"?: SwirlBoxPadding1;
+        "paddingBlockStart"?: SwirlBoxPadding1;
+        "paddingInlineEnd"?: SwirlBoxPadding1;
+        "paddingInlineStart"?: SwirlBoxPadding1;
         "spacing"?: SwirlStackSpacing;
     }
     interface SwirlResourceListFileItem {
@@ -2211,7 +2220,10 @@ export namespace Components {
     }
     interface SwirlTable {
         "caption"?: string;
+        "dragDropHandle"?: string;
+        "dragDropInstructions": { end: string; initial: string; moved: string; start: string; };
         "emptyStateLabel"?: string;
+        "enableDragDrop"?: boolean;
         "label": string;
         /**
           * Force a re-render of the table
@@ -2641,6 +2653,10 @@ export interface SwirlSwitchCustomEvent<T> extends CustomEvent<T> {
 export interface SwirlTabBarCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLSwirlTabBarElement;
+}
+export interface SwirlTableCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLSwirlTableElement;
 }
 export interface SwirlTabsCustomEvent<T> extends CustomEvent<T> {
     detail: T;
@@ -5299,7 +5315,18 @@ declare global {
         prototype: HTMLSwirlTabBarElement;
         new (): HTMLSwirlTabBarElement;
     };
+    interface HTMLSwirlTableElementEventMap {
+        "dropRow": SwirlTableDropRowEvent;
+    }
     interface HTMLSwirlTableElement extends Components.SwirlTable, HTMLStencilElement {
+        addEventListener<K extends keyof HTMLSwirlTableElementEventMap>(type: K, listener: (this: HTMLSwirlTableElement, ev: SwirlTableCustomEvent<HTMLSwirlTableElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLSwirlTableElementEventMap>(type: K, listener: (this: HTMLSwirlTableElement, ev: SwirlTableCustomEvent<HTMLSwirlTableElementEventMap[K]>) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
     }
     var HTMLSwirlTableElement: {
         prototype: HTMLSwirlTableElement;
@@ -7510,6 +7537,11 @@ declare namespace LocalJSX {
     oldIndex: number;
     newIndex: number;
   }>) => void;
+        "padding"?: SwirlBoxPadding1;
+        "paddingBlockEnd"?: SwirlBoxPadding1;
+        "paddingBlockStart"?: SwirlBoxPadding1;
+        "paddingInlineEnd"?: SwirlBoxPadding1;
+        "paddingInlineStart"?: SwirlBoxPadding1;
         "spacing"?: SwirlStackSpacing;
     }
     interface SwirlResourceListFileItem {
@@ -7877,8 +7909,12 @@ declare namespace LocalJSX {
     }
     interface SwirlTable {
         "caption"?: string;
+        "dragDropHandle"?: string;
+        "dragDropInstructions"?: { end: string; initial: string; moved: string; start: string; };
         "emptyStateLabel"?: string;
+        "enableDragDrop"?: boolean;
         "label": string;
+        "onDropRow"?: (event: SwirlTableCustomEvent<SwirlTableDropRowEvent>) => void;
     }
     interface SwirlTableCell {
     }