+++ packages/swirl-components/src/components/swirl-table/swirl-table.spec.tsx
@@ -29,44 +29,40 @@ describe("swirl-table", () => {
     });
 
     expect(page.root).toEqualHtml(`
-      <swirl-table caption="Caption" label="Label">
-        <mock:shadow-root>
-          <div class="table">
-            <div class="table__container">
-              <div aria-describedby="caption" aria-label="Label" class="table__table" role="table">
-                <swirl-visually-hidden>
-                  <div id="caption">
-                    Caption
-                  </div>
-                </swirl-visually-hidden>
-                <div role="rowgroup">
-                  <div class="table__header" role="row">
-                    <slot name="columns"></slot>
+      <swirl-table caption=\"Caption\" label=\"Label\">
+        <div class=\"table\">
+          <div class=\"table__container\" tabIndex=\"-1\">
+            <div aria-describedby=\"caption\" aria-label=\"Label\" class=\"table__table\" role=\"table\">
+              <swirl-visually-hidden>
+                <div id=\"caption\">
+                  Caption
+                </div>
+              </swirl-visually-hidden>
+              <div role=\"rowgroup\">
+                <div class=\"table__header\" role=\"row\">
+                  <div slot=\"columns\">
+                    <span role=\"columnheader\">
+                      Column 1
+                    </span>
+                    <span role=\"columnheader\">
+                      Column 2
+                    </span>
                   </div>
                 </div>
-                <div class="table__body">
-                  <slot name="rows"></slot>
+              </div>
+              <div class=\"table__body\">
+                <div slot=\"rows\">
+                  <div role=\"row\">
+                    <span role=\"cell\">
+                      1
+                    </span>
+                    <span role=\"cell\">
+                      2
+                    </span>
+                  </div>
                 </div>
               </div>
             </div>
-          </div>
-        </mock:shadow-root>
-        <div slot="columns">
-          <span role="columnheader">
-            Column 1
-          </span>
-          <span role="columnheader">
-            Column 2
-          </span>
-        </div>
-        <div slot="rows">
-          <div role="row">
-            <span role="cell">
-              1
-            </span>
-            <span role="cell">
-              2
-            </span>
           </div>
         </div>
       </swirl-table>